Fix incorrect CSS escaping (fixes #188)
The problem wasn't in what was being done. The problem is that bleach will
automatically HTML-escape its output. That's fine, but it means there's no way
to use it with CSS. Instead, I'm base64 encoding the CSS as a <link>, which
will prevent issues with XSS by relegating the bad content to a CSS file.